pyproject.toml
[project]
name = "personal-backend-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3 ~= 1.35.39",
    "boto3-stubs[essential] ~= 1.35.39",
    "fastapi ~= 0.115.1",
    "httpx ~= 0.27.0",
    "mangum ~= 0.19.0",
    "markdown ~= 3.7",
    "pendulum ~= 3.0.0",
    "pydantic ~= 2.9.2",
    "pydantic-settings ~= 2.5.2",
    "pyjwt ~= 2.9.0",
    "python-slugify ~= 8.0.4",
    "sshfs>=2025.2.0",
    "unidecode ~= 1.3.8",
    "uvicorn ~= 0.31.1",
]

[dependency-groups]
dev = [
    "autoflake ~= 2.3.1",
    "aws-lambda-powertools[validation] ~= 3.10.0",
    "bandit ~= 1.8.3",
    "black ~= 24.10.0",
    "faker ~= 30.3.0",
    "isort ~= 5.13.2",
    "moto ~= 5.0.17",
    "mypy ~= 1.12.0",
    "pycodestyle ~= 2.12.1",
    "pytest ~= 8.3.3",
    "pytest-cov ~= 5.0.0",
    "pytest-env ~= 1.1.5",
    "pytest-mock ~= 3.14.0",
    "pytest-xdist ~= 3.6.1",
    "respx ~= 0.22.0",
]
